requirements.txt
django==3.2.6
djangorestframework~=3.12.4
requests~=2.26.0
jwt~=1.2.0